+++ .gitignore
@@ -2,6 +2,7 @@
 
 # dependencies
 /node_modules
+/src/pages/TabBarNav
 
 # testing
 /coverage
