+++ src/index.js
@@ -14,6 +14,7 @@ import './style/index.css'
 import Home from './pages/Home/Index'
 import Detail from './pages/Detail/Index'
 import Search from './pages/Search/Index'
+import List from './pages/List/Index'
 import registerServiceWorker from './registerServiceWorker'
 
 // history 模式
@@ -25,6 +26,7 @@ ReactDOM.render(
       <Route exact path="/" component={Home} />
       <Route exact path="/detail/:id" component={Detail} />
       <Route exact path="/search" component={Search} />
+      <Route exact path="/list" component={List} />
     </Switch>
   </Router>,
   document.getElementById('App')
