+++ src/style/Home.scss
@@ -190,66 +190,9 @@
   }
 }
 
-
-.Home-content-all {
-  display: flex;
-  margin-top: 0;
-}
-
-.Sidebar {
-  display: flex;
-  width: 160px;
-  flex-direction: column;
-  padding: 0;
-  margin-right: 40px;
-  white-space: pre;
-  overflow-x: hidden;
-  // background-color: #f5f5f5;
-
-  a {
-    color: #383838;
-    padding: 20px 0;
-    text-align: center;
-    transition: all 0.2s linear;
-
-    &.actived {
-      position: relative;
-      background-color: #f5f5f5;
-
-      &::after {
-        content: '';
-        position: absolute;
-        left: 0;
-        top: 0;
-        bottom: 0;
-        width: 10px;
-        background-color: #108ee9;
-      }
-    }
-  }
-}
-
-.Home-content-all-cook {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  flex: 1;
-
-  .cook {
-    margin-right: 10px;
-    width: 150px;
-    overflow: hidden;
-  }
-
-  .cover {
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position: center;
-    width: 100%;
-    height: 150px;
-  }
-
-  .title {
-    text-align: center;
-  }
+.Tabbar {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  width: 100%;
 }
