+++ src/style/index.css
@@ -1,6 +1,7 @@
 @import url(./Home.css);
 @import url(./Detail.css);
 @import url(./Search.css);
+@import url(./List.css);
 html,
 body,
 #App {
@@ -13,7 +14,7 @@ body {
   color: #383838;
   font-size: 40px;
   line-height: 1.5;
-  font-family: "PingFang SC", "Helvetica Neue", Helvetica,Arial, "Microsoft Yahei", "Hiragino Sans GB", "Heiti SC", "WenQuanYi Micro Hei", sans-serif;
+  font-family: "PingFang SC", "Helvetica Neue", Helvetica,Arial, "Microsoft Yahei", "Hiragino Sans GB", "Heiti SC", "WenQuanYi Micro Hei", sans-serif;
   width: 100%;
   overflow-x: hidden; }
