+++ src/style/index.scss
@@ -30,3 +30,4 @@ body {
 @import './Home.css';
 @import './Detail.css';
 @import './Search.css';
+@import './List.css'
